+++ src/services/user.js
@@ -10,6 +10,18 @@ exports.findBasedOnQuery = (data) => User.find(data);
 
 exports.findUserById = (id) => User.findById(id);
 
+exports.findUsersByIds = async (ids) => {
+  const result = [];
+  for (const each of ids) {
+    const user = await User.findById(each);
+
+    delete user.password;
+    result.push(user);
+  }
+
+  return result;
+};
+
 exports.findAllUsers = (query = {}) => User.find(query);
 
 exports.removeUser = (id) => User.findByIdAndRemove(id);
@@ -22,7 +34,7 @@ exports.deleteUserByEmail = (email) => User.deleteOne({ email });
 exports.search = async ({ searchInput, id }) => {
   const user = await User.find(
     { $text: { $search: searchInput }, company: id },
-    { score: { $meta: 'textScore' } },
+    { score: { $meta: 'textScore' } }
   ).sort({ score: { $meta: 'textScore' } });
 
   return user;